test: 🔄 projectFolderIgnoreList: move to separate file
